requirements.txt
# Bill Software Dependencies
Pillow>=9.0.0
matplotlib>=3.5.0
pyinstaller>=5.0.0
cairosvg>=2.5.0